+++ python/oemaestro/__init__.py
@@ -27,8 +27,8 @@ import warnings
 from importlib import metadata
 from pathlib import Path
 
-__version__ = "0.8.1"
-__version_info__ = (0, 8, 1)
+__version__ = "0.8.2"
+__version_info__ = (0, 8, 2)
 
 
 _OPENEYE_COMPAT_PRELOAD_PATHS: list[str] = []